cohfar: change log10(cohsnr) and log10(chisq) to be double presicion; change snr ticks to be 0.6:0.01:2.49; chisq ticks to be -0.5:0.01:2.99

+++ gstlal-ugly/gst/cuda/cohfar/cohfar_accumbackground.c
@@ -171,7 +171,7 @@ static GstFlowReturn cohfar_accumbackground_chain(GstPad *pad, GstBuffer *inbuf)
 		if (intable->is_background == 1) {
 			//printf("cohsnr %f, maxsnr %f\n", intable->cohsnr, intable->maxsnglsnr);
 			icombo = get_icombo(intable->ifos);
-			background_stats_rates_update(intable->cohsnr, intable->chisq, stats[icombo]->rates);
+			background_stats_rates_update((double)intable->cohsnr, (double)intable->chisq, stats[icombo]->rates);
 		} else { /* coherent trigger entry */
 			memcpy(outtable, intable, sizeof(PostcohInspiralTable));
 			outtable++;
